% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examp}
\alias{examp}
\title{List of the different outputs of the main function for examples}
\format{
A list with 4 components:
\describe{
\item{dataProfile}{Output of the profileGLMM_preprocess() function example}
\item{MCMC_Obj}{Output of the profileGLMM_Gibbs() function example}
\item{post_Obj}{Output of the profileGLMM_postprocess() function example}
\item{pred_Obj}{Output of the profileGLMM_predict() function example}
}
}
\source{
Generated synthetically by the package authors.
}
\usage{
examp
}
\description{
A list of the different outputs of the main function for examples
}
\keyword{datasets}
