% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm_fit-methods.R
\name{plot.pglmm_fit}
\alias{plot.pglmm_fit}
\title{Plot method for pglmm_fit continuous covariates cluster characteristics}
\usage{
\method{plot}{pglmm_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pglmm_fit}}

\item{...}{Additional arguments\itemize{
\item title : main title of the plot
\item color : palette to be used }}
}
\description{
Plot method for pglmm_fit continuous covariates cluster characteristics
}
