% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm_data-methods.R
\name{print.pglmm_data}
\alias{print.pglmm_data}
\title{Print method for pglmm_data}
\usage{
\method{print}{pglmm_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pglmm_data}}

\item{...}{Additional arguments}
}
\description{
Print method for pglmm_data
}
