% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm_fit-methods.R
\name{print.pglmm_fit}
\alias{print.pglmm_fit}
\title{Print method for pglmm_fit}
\usage{
\method{print}{pglmm_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pglmm_fit}}

\item{...}{Additional arguments}
}
\description{
Print method for pglmm_fit
}
