% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm_mcmc-methods.R
\name{print.pglmm_mcmc}
\alias{print.pglmm_mcmc}
\title{Print method for pglmm_mcmc}
\usage{
\method{print}{pglmm_mcmc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pglmm_mcmc}}

\item{...}{Additional arguments}
}
\description{
Print method for pglmm_mcmc
}
