% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileGLMM_preprocess.R
\name{profileGLMM_preprocess}
\alias{profileGLMM_preprocess}
\title{Preprocess the data from a list describing the profile LMM model}
\usage{
profileGLMM_preprocess(
  covList,
  dataframe,
  nC,
  intercept = list(FE = T, RE = T, Lat = T)
)
}
\arguments{
\item{covList}{A list with fields:\itemize{
\item  FE fixed effect covariates names/index in dataframe
\item  RE random effect covariates names/index in dataframe
\item  Lat latent effect covariates names/index in dataframe
\item  Assign assignement variable categorical not supported yet
\item  REunit statistical unit of the RE colomn name/index
\item  Y outcome (Continuous)}}

\item{dataframe}{A dataframe containing outcome anf covariates}

\item{nC}{int: maximal number of cluster for the DP truncation}

\item{intercept}{(optionnal): A list with fields\itemize{
\item  RE bool indicating if FE have an intercept
\item  FE bool indicating if RE have an intercept
\item  Lat bool indicating if Latent have an intercept}}
}
\value{
A list with\itemize{
\item  d dictionary with  [XFE,XRE,XLat,U,ZRE] design matrices
\item  [[params]] list of the parameters of the data\itemize{
                 \item  n int nb of obs
                 \item  qFE lint, number of covariates of FE
                 \item  nRE int, number of stat units of RE
                 \item  qRE int, number of covariates of RE}
\item prior a list with all the specification of the default prior used
\item theta a list with a default set of parameters to start the chain, drawn from the prior}
}
\description{
Preprocess the data from a list describing the profile LMM model
}
