% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_init.R
\name{theta_init}
\alias{theta_init}
\title{Initatilize the variables for the gibbs sampler.}
\usage{
theta_init(prior, params, nC)
}
\arguments{
\item{prior}{A list of prior configuration to draw initialization from.
Same structure as the output of prior_init.}

\item{params}{A list of the problem parameters.
Same structure as the output of process_Data_outcome.}

\item{nC}{number of possible clusters}
}
\value{
A list containing the initalisation values of the Gibbs sampler
}
\description{
Initatilize the variables for the gibbs sampler.
}
