% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeCat.R
\name{encodeCat}
\alias{encodeCat}
\title{Encodes factor variables dummies LAST is reference}
\usage{
encodeCat(dataframe)
}
\arguments{
\item{dataframe}{dataframe to convert}
}
\value{
A dataframe with factor variable dummy encoded
}
\description{
Encodes factor variables dummies LAST is reference
}
