% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm_fit-methods.R
\name{predict.pglmm_fit}
\alias{predict.pglmm_fit}
\title{Prediction of cluster memberships and outcomes}
\usage{
\method{predict}{pglmm_fit}(x, newData, ...)
}
\arguments{
\item{newData}{: A list with fields\itemize{
\item XFE A numeric matrix of fixed effects covariates for the prediction data.
\item XLat A numeric matrix of latent effect covariates.
\item UCont A numeric matrix or vector of continuous profile variables. Defaults to \code{NULL}.
\item UCat A numeric matrix or vector of categorical profile variables. Defaults to \code{NULL}.}}

\item{...}{Additional arguments}

\item{object}{An object of class \code{pglmm_fit} .}
}
\description{
(This documentation is now for internal use only)
}
\examples{
# Load post_Obj, the result of profileGLMM_postProcess()
data("examp")
post_Obj = examp$post_Obj

# run prediction for training data
pred_Obj = predict(post_Obj,examp$dataProfile)


}
