% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_init.R
\name{prior_init}
\alias{prior_init}
\title{Initialise the priors of the model}
\usage{
prior_init(params, nC)
}
\arguments{
\item{params}{A list with Fields: \itemize{
\item FE fixed effect covariates names/index in dataframe
\item RE random effect covariates names/index in dataframe
\item REunit statistical unit of the RE colomn name/index
\item Lat latent effect covariates names/index in dataframe
\item Assign assignement variable categorical not supported yet}}

\item{nC}{An integer, the number of possible clusters}
}
\value{
A list of the priors parameters for each variable
}
\description{
Initialise the priors of the model
}
