% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileGLMM_Gibbs.R
\name{profileGLMM_Gibbs}
\alias{profileGLMM_Gibbs}
\title{R wrapper to call the cpp GSLoopCPP function.}
\usage{
profileGLMM_Gibbs(proLMMObj, nIt, nBurnIn)
}
\arguments{
\item{proLMMObj}{Profile LMM  preprocessed by the process_Data_outcome function}

\item{nIt}{Integer, total number of elements sampled by the gibbs sampler}

\item{nBurnIn}{Integer, number of elements of the chain that are not saved as part of the burn-in.}
}
\value{
The Gibbs-sampled posterior. Output of the cpp GSLoopCPP function.
}
\description{
R wrapper to call the cpp GSLoopCPP function.
}
